🚨 EMERGENCY HOTFIX: Fix CI failures - rustfmt and clippy issues
## Emergency Fix for Production Issue

### Impact
- **Severity**: P1
- **Affected Systems**: CI/CD Pipeline
- **User Impact**: Development workflow blocked

### Root Cause
Three formatting and linting issues introduced in recent f64 migration:
1. Extra blank line in action_test.rs causing rustfmt failure
2. Unnecessary f64 cast in game.rs causing clippy failure
3. Non-inlined format args in joker_registry.rs causing clippy failure

### Solution
Applied minimal fixes:
- Fixed formatting issue in core/tests/action_test.rs:177
- Removed unnecessary f64 cast in core/src/game.rs:684
- Updated format string to use inline args in core/src/joker_registry.rs:132

### Testing
- [x] cargo fmt --check passes
- [x] cargo clippy --all -- -D warnings passes
- [x] cargo test --all passes (253 tests)
- [x] Manual verification completed

+++ core/src/game.rs
@@ -681,7 +681,7 @@ impl Game {
     }
 
     fn calc_reward(&mut self, blind: Blind) -> Result<f64, GameError> {
-        let mut interest = (self.money * self.config.interest_rate as f64).floor();
+        let mut interest = (self.money * self.config.interest_rate).floor();
         if interest > self.config.interest_max as f64 {
             interest = self.config.interest_max as f64
         }
